+++ e2e/e2etests/test_erc20_deposit_restricted_address.go
@@ -1,8 +1,6 @@
 package e2etests
 
 import (
-	"math/big"
-
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/require"
 
@@ -13,8 +11,8 @@ import (
 func TestERC20DepositRestricted(r *runner.E2ERunner, args []string) {
 	require.Len(r, args, 1)
 
-	amount, ok := big.NewInt(0).SetString(args[0], 10)
-	require.True(r, ok)
+	// parse the deposit amount
+	amount := parseBigInt(r, args[0])
 
 	// deposit ERC20 to restricted address
 	r.DepositERC20WithAmountAndMessage(ethcommon.HexToAddress(testutils.RestrictedEVMAddressTest), amount, []byte{})