+++ zetaclient/chains/evm/signer/signer_test.go
@@ -10,7 +10,6 @@ import (
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/rs/zerolog"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	observertypes "github.com/zeta-chain/zetacore/x/observer/types"
 	zctx "github.com/zeta-chain/zetacore/zetaclient/context"
@@ -45,7 +44,7 @@ func getNewEvmSigner(tss interfaces.TSSSigner) (*Signer, error) {
 		tss = mocks.NewTSSMainnet()
 	}
 
-	mpiAddress := ConnectorAddress
+	connectorAddress := ConnectorAddress
 	erc20CustodyAddress := ERC20CustodyAddress
 	logger := base.Logger{}
 
@@ -56,10 +55,9 @@ func getNewEvmSigner(tss interfaces.TSSSigner) (*Signer, error) {
 		nil,
 		logger,
 		mocks.EVMRPCEnabled,
-		config.GetConnectorABI(),
-		config.GetERC20CustodyABI(),
-		mpiAddress,
+		connectorAddress,
 		erc20CustodyAddress,
+		sample.EthAddress(),
 	)
 }
 
@@ -184,232 +182,6 @@ func TestSigner_TryProcessOutbound(t *testing.T) {
 	require.Len(t, *list, 1)
 }
 
-func TestSigner_SignOutbound(t *testing.T) {
-	ctx := makeCtx(t)
-
-	// Setup evm signer
-	tss := mocks.NewTSSMainnet()
-	evmSigner, err := getNewEvmSigner(tss)
-	require.NoError(t, err)
-
-	// Setup txData struct
-
-	cctx := getCCTX(t)
-	txData, skip, err := NewOutboundData(ctx, cctx, 123, zerolog.Logger{})
-	require.False(t, skip)
-	require.NoError(t, err)
-
-	t.Run("SignOutbound - should successfully sign LegacyTx", func(t *testing.T) {
-		// Call SignOutbound
-		tx, err := evmSigner.SignOutbound(ctx, txData)
-		require.NoError(t, err)
-
-		// Verify Signature
-		tss := mocks.NewTSSMainnet()
-		verifyTxSignature(t, tx, tss.Pubkey(), evmSigner.EvmSigner())
-
-		// check that by default tx type is legacy tx
-		assert.Equal(t, ethtypes.LegacyTxType, int(tx.Type()))
-	})
-	t.Run("SignOutbound - should fail if keysign fails", func(t *testing.T) {
-		// Pause tss to make keysign fail
-		tss.Pause()
-
-		// Call SignOutbound
-		tx, err := evmSigner.SignOutbound(ctx, txData)
-		require.ErrorContains(t, err, "sign onReceive error")
-		require.Nil(t, tx)
-	})
-
-	t.Run("SignOutbound - should successfully sign DynamicFeeTx", func(t *testing.T) {
-		// ARRANGE
-		const (
-			gwei        = 1_000_000_000
-			priorityFee = 1 * gwei
-			gasPrice    = 3 * gwei
-		)
-
-		// Given a CCTX with gas price and priority fee
-		cctx := getCCTX(t)
-		cctx.OutboundParams[0].GasPrice = big.NewInt(gasPrice).String()
-		cctx.OutboundParams[0].GasPriorityFee = big.NewInt(priorityFee).String()
-
-		// Given outbound data
-		txData, skip, err := NewOutboundData(ctx, cctx, 123, makeLogger(t))
-		require.False(t, skip)
-		require.NoError(t, err)
-
-		// Given a working TSS
-		tss.Unpause()
-
-		// ACT
-		tx, err := evmSigner.SignOutbound(ctx, txData)
-		require.NoError(t, err)
-
-		// ASSERT
-		verifyTxSignature(t, tx, mocks.NewTSSMainnet().Pubkey(), evmSigner.EvmSigner())
-
-		// check that by default tx type is a dynamic fee tx
-		assert.Equal(t, ethtypes.DynamicFeeTxType, int(tx.Type()))
-
-		// check that the gasPrice & priorityFee are set correctly
-		assert.Equal(t, int64(gasPrice), tx.GasFeeCap().Int64())
-		assert.Equal(t, int64(priorityFee), tx.GasTipCap().Int64())
-	})
-}
-
-func TestSigner_SignRevertTx(t *testing.T) {
-	ctx := makeCtx(t)
-
-	// Setup evm signer
-	tss := mocks.NewTSSMainnet()
-	evmSigner, err := getNewEvmSigner(tss)
-	require.NoError(t, err)
-
-	// Setup txData struct
-	cctx := getCCTX(t)
-	txData, skip, err := NewOutboundData(ctx, cctx, 123, zerolog.Logger{})
-	require.False(t, skip)
-	require.NoError(t, err)
-
-	t.Run("SignRevertTx - should successfully sign", func(t *testing.T) {
-		// Call SignRevertTx
-		tx, err := evmSigner.SignRevertTx(ctx, txData)
-		require.NoError(t, err)
-
-		// Verify tx signature
-		tss := mocks.NewTSSMainnet()
-		verifyTxSignature(t, tx, tss.Pubkey(), evmSigner.EvmSigner())
-
-		// Verify tx body basics
-		// Note: Revert tx calls connector contract with 0 gas token
-		verifyTxBodyBasics(t, tx, evmSigner.zetaConnectorAddress, txData.nonce, big.NewInt(0))
-	})
-	t.Run("SignRevertTx - should fail if keysign fails", func(t *testing.T) {
-		// Pause tss to make keysign fail
-		tss.Pause()
-
-		// Call SignRevertTx
-		tx, err := evmSigner.SignRevertTx(ctx, txData)
-		require.ErrorContains(t, err, "sign onRevert error")
-		require.Nil(t, tx)
-	})
-}
-
-func TestSigner_SignCancelTx(t *testing.T) {
-	ctx := makeCtx(t)
-
-	// Setup evm signer
-	tss := mocks.NewTSSMainnet()
-	evmSigner, err := getNewEvmSigner(tss)
-	require.NoError(t, err)
-
-	// Setup txData struct
-	cctx := getCCTX(t)
-	txData, skip, err := NewOutboundData(ctx, cctx, 123, zerolog.Logger{})
-	require.False(t, skip)
-	require.NoError(t, err)
-
-	t.Run("SignCancelTx - should successfully sign", func(t *testing.T) {
-		// Call SignRevertTx
-		tx, err := evmSigner.SignCancelTx(ctx, txData)
-		require.NoError(t, err)
-
-		// Verify tx signature
-		tss := mocks.NewTSSMainnet()
-		verifyTxSignature(t, tx, tss.Pubkey(), evmSigner.EvmSigner())
-
-		// Verify tx body basics
-		// Note: Cancel tx sends 0 gas token to TSS self address
-		verifyTxBodyBasics(t, tx, evmSigner.TSS().EVMAddress(), txData.nonce, big.NewInt(0))
-	})
-	t.Run("SignCancelTx - should fail if keysign fails", func(t *testing.T) {
-		// Pause tss to make keysign fail
-		tss.Pause()
-
-		// Call SignCancelTx
-		tx, err := evmSigner.SignCancelTx(ctx, txData)
-		require.ErrorContains(t, err, "SignCancelTx error")
-		require.Nil(t, tx)
-	})
-}
-
-func TestSigner_SignWithdrawTx(t *testing.T) {
-	ctx := makeCtx(t)
-
-	// Setup evm signer
-	tss := mocks.NewTSSMainnet()
-	evmSigner, err := getNewEvmSigner(tss)
-	require.NoError(t, err)
-
-	// Setup txData struct
-	cctx := getCCTX(t)
-	txData, skip, err := NewOutboundData(ctx, cctx, 123, zerolog.Logger{})
-	require.False(t, skip)
-	require.NoError(t, err)
-
-	t.Run("SignWithdrawTx - should successfully sign", func(t *testing.T) {
-		// Call SignWithdrawTx
-		tx, err := evmSigner.SignWithdrawTx(ctx, txData)
-		require.NoError(t, err)
-
-		// Verify tx signature
-		tss := mocks.NewTSSMainnet()
-		verifyTxSignature(t, tx, tss.Pubkey(), evmSigner.EvmSigner())
-
-		// Verify tx body basics
-		verifyTxBodyBasics(t, tx, txData.to, txData.nonce, txData.amount)
-	})
-	t.Run("SignWithdrawTx - should fail if keysign fails", func(t *testing.T) {
-		// Pause tss to make keysign fail
-		tss.Pause()
-
-		// Call SignWithdrawTx
-		tx, err := evmSigner.SignWithdrawTx(ctx, txData)
-		require.ErrorContains(t, err, "SignWithdrawTx error")
-		require.Nil(t, tx)
-	})
-}
-
-func TestSigner_SignERC20WithdrawTx(t *testing.T) {
-	ctx := makeCtx(t)
-
-	// Setup evm signer
-	tss := mocks.NewTSSMainnet()
-	evmSigner, err := getNewEvmSigner(tss)
-	require.NoError(t, err)
-
-	// Setup txData struct
-	cctx := getCCTX(t)
-	txData, skip, err := NewOutboundData(ctx, cctx, 123, zerolog.Logger{})
-	require.False(t, skip)
-	require.NoError(t, err)
-
-	t.Run("SignERC20WithdrawTx - should successfully sign", func(t *testing.T) {
-		// Call SignERC20WithdrawTx
-		tx, err := evmSigner.SignERC20WithdrawTx(ctx, txData)
-		require.NoError(t, err)
-
-		// Verify tx signature
-		tss := mocks.NewTSSMainnet()
-		verifyTxSignature(t, tx, tss.Pubkey(), evmSigner.EvmSigner())
-
-		// Verify tx body basics
-		// Note: Withdraw tx calls erc20 custody contract with 0 gas token
-		verifyTxBodyBasics(t, tx, evmSigner.er20CustodyAddress, txData.nonce, big.NewInt(0))
-	})
-
-	t.Run("SignERC20WithdrawTx - should fail if keysign fails", func(t *testing.T) {
-		// pause tss to make keysign fail
-		tss.Pause()
-
-		// Call SignERC20WithdrawTx
-		tx, err := evmSigner.SignERC20WithdrawTx(ctx, txData)
-		require.ErrorContains(t, err, "sign withdraw error")
-		require.Nil(t, tx)
-	})
-}
-
 func TestSigner_BroadcastOutbound(t *testing.T) {
 	ctx := makeCtx(t)
 
@@ -424,8 +196,8 @@ func TestSigner_BroadcastOutbound(t *testing.T) {
 	require.False(t, skip)
 
 	t.Run("BroadcastOutbound - should successfully broadcast", func(t *testing.T) {
-		// Call SignERC20WithdrawTx
-		tx, err := evmSigner.SignERC20WithdrawTx(ctx, txData)
+		// Call SignERC20Withdraw
+		tx, err := evmSigner.SignERC20Withdraw(ctx, txData)
 		require.NoError(t, err)
 
 		evmSigner.BroadcastOutbound(
@@ -463,7 +235,7 @@ func TestSigner_SignerErrorMsg(t *testing.T) {
 }
 
 func makeCtx(t *testing.T) context.Context {
-	app := zctx.New(config.New(false), zerolog.Nop())
+	app := zctx.New(config.New(false), nil, zerolog.Nop())
 
 	bscParams := mocks.MockChainParams(chains.BscMainnet.ChainId, 10)