+++ zetaclient/zetacore/constant.go
@@ -16,7 +16,7 @@ const (
 	PostVoteInboundGasLimit = 500_000
 
 	// PostVoteInboundExecutionGasLimit is the gas limit for voting on observed inbound tx and executing it
-	PostVoteInboundExecutionGasLimit = 4_000_000
+	PostVoteInboundExecutionGasLimit = 6_500_000
 
 	// PostVoteInboundMessagePassingExecutionGasLimit is the gas limit for voting on, and executing ,observed inbound tx related to message passing (coin_type == zeta)
 	PostVoteInboundMessagePassingExecutionGasLimit = 4_000_000
@@ -41,7 +41,7 @@ const (
 
 	// PostVoteOutboundRevertGasLimit is the gas limit for voting on observed outbound tx for revert (when outbound fails)
 	// The value needs to be higher because reverting implies interacting with the EVM to perform swaps for the gas token
-	PostVoteOutboundRevertGasLimit = 1_500_000
+	PostVoteOutboundRevertGasLimit = 4_000_000
 )
 
 // constants for monitoring tx results